Fix filter construction bugs
FilterFactory._build_freq_response() had a feature to mask the FFT
coefficients of the measured response outside of a specified band.  This
is sensible since it lets us avoid massive gain to extend the bandwidth
of a driver.  However, this function didn't consider that the frequency
scale wraps around and the frequency response (of the real impulse
response) needs to be conjugate-symmetric.

This commit adds a new function _mask_freq_response() to do the masking
properly.  Now there's no need to try oddball stuff like taking the 4th
power of FFT coefficients.

Hopefully Mike will verify this technique with measurements.